+++ Assets/Scripts/Character/Player/PlayerState.cs
@@ -1,6 +1,6 @@
 public class PlayerState : BaseState
 {
-    public PlayerState(float speed, float health, float maxHealth, float bonus) : base(speed, health, maxHealth, bonus)
+    public PlayerState(float speed, int health, int maxHealth, int bonus) : base(speed, health, maxHealth, bonus)
     {
     }
 
@@ -10,9 +10,9 @@ public class PlayerState : BaseState
 
     public override BaseState Mutate(
         float? speed = null,
-        float? health = null,
-        float? maxHealth = null,
-        float? bonus = null)
+        int? health = null,
+        int? maxHealth = null,
+        int? bonus = null)
     {
         return new PlayerState(
             speed ?? Speed,
