+++ Assets/Scripts/Character/State/BaseState.cs
@@ -5,16 +5,16 @@ using UnityEngine;
 public class BaseState
 {
     public float Speed { get; }
-    public float Health { get; }
-    public float MaxHealth { get; }
-    public float Bonus { get; }
+    public int Health { get; }
+    public int MaxHealth { get; }
+    public int Bonus { get; }
 
-    public BaseState(float speed, float health, float maxHealth, float bonus)
+    public BaseState(float speed, int health, int maxHealth, int bonus)
     {
         Speed = Mathf.Clamp(speed, 0, Mathf.Infinity);
         MaxHealth = maxHealth;
         Health = Mathf.Clamp(health, 0, maxHealth);
-        Bonus = Mathf.Clamp(bonus, 0, Mathf.Infinity);
+        Bonus = (int)Mathf.Clamp(bonus, 0, Mathf.Infinity);
     }
 
     public BaseState(CharacterStarterInfo starerInfo) : this(starerInfo.Speed, starerInfo.MaxHealth, starerInfo.MaxHealth, starerInfo.Bonus)
@@ -24,9 +24,9 @@ public class BaseState
 
     public virtual BaseState Mutate(
         float? speed = null,
-        float? health = null,
-        float? maxHealth = null,
-        float? bonus = null
+        int? health = null,
+        int? maxHealth = null,
+        int? bonus = null
     )
     {
         return new BaseState(
